requirements.txt
python-socketio
picamera2>=0.3.12
opencv-python-headless>=4.8.0
numpy>=1.24.0
google-generativeai>=0.3.0

